+++ core/model/rules.go
@@ -44,6 +44,13 @@ type SymptomRule struct {
 	Items []SymptomRuleItem
 }
 
+//CRules represents all rules in a raw format
+type CRules struct {
+	AppVersion string `json:"app_version" bson:"app_version"`
+	CountyID   string `json:"county_id" bson:"county_id"`
+	Data       string `json:"data" bson:"data"`
+}
+
 //SymptomRuleItem represents a symptom rule item entity
 type SymptomRuleItem struct {
 	Gr1          bool