+++ core/services.go
@@ -333,6 +333,19 @@ func (app *Application) getSymptomRuleByCounty(countyID string) (*model.SymptomR
 	return symptomRule, countyStatuses, nil
 }
 
+func (app *Application) getCRulesByCounty(appVersion *string, countyID string) (*model.CRules, error) {
+	v, err := app.checkAppVersion(appVersion)
+	if err != nil {
+		return nil, err
+	}
+
+	rules, err := app.storage.FindCRulesByCountyID(*v, countyID)
+	if err != nil {
+		return nil, err
+	}
+	return rules, nil
+}
+
 func (app *Application) getAccessRuleByCounty(countyID string) (*model.AccessRule, []*model.CountyStatus, error) {
 	//get the access rule
 	accessRule, err := app.storage.FindAccessRuleByCountyID(countyID)
@@ -407,4 +420,17 @@ func (app *Application) updateEHistory(userID string, ID string, date *time.Time
 	}
 
 	return history, nil
+}
+
+func (app *Application) getSymptoms(appVersion *string) (*model.Symptoms, error) {
+	v, err := app.checkAppVersion(appVersion)
+	if err != nil {
+		return nil, err
+	}
+
+	symptoms, err := app.storage.ReadSymptoms(*v)
+	if err != nil {
+		return nil, err
+	}
+	return symptoms, nil
 }