+++ driver/web/rest/adminapis.go
@@ -82,6 +82,34 @@ func (h AdminApisHandler) UpdateCovid19Config(current model.User, group string, 
 	w.Write([]byte("Successfully updated"))
 }
 
+//GetAppVersions gives the supported app versions
+// @Description Gives the supported app versions
+// @Tags Admin
+// @ID GetAppVersions
+// @Accept  json
+// @Success 200 {array} string
+// @Security AdminUserAuth
+// @Security AdminGroupAuth
+// @Router /admin/app-versions [get]
+func (h AdminApisHandler) GetAppVersions(current model.User, group string, w http.ResponseWriter, r *http.Request) {
+	appVersions, err := h.app.Administration.GetAppVersions()
+	if err != nil {
+		log.Println("Error on getting the app versions")
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+	data, err := json.Marshal(appVersions)
+	if err != nil {
+		log.Println("Error on marshal the app versions")
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(data)
+}
+
 //GetNews gets news
 // @Description Gives news.
 // @Tags Admin
@@ -1974,19 +2002,13 @@ func (h AdminApisHandler) GetTestTypes(current model.User, group string, w http.
 }
 
 type createTestTypeResultRequest struct {
-	TestTypeID          string `json:"test_type_id" validate:"uuid"`
-	Name                string `json:"name" validate:"required"`
-	NextStep            string `json:"next_step" validate:"required"`
-	NextStepOffset      *int   `json:"next_step_offset"`
-	ResultExpiresOffset *int   `json:"result_expires_offset"`
+	TestTypeID string `json:"test_type_id" validate:"uuid"`
+	Name       string `json:"name" validate:"required"`
 } //@name createTestTypeResultRequest
 
 type createTestTypeResultResponse struct {
-	ID                  string `json:"id"`
-	Name                string `json:"name"`
-	NextStep            string `json:"next_step"`
-	NextStepOffset      *int   `json:"next_step_offset"`
-	ResultExpiresOffset *int   `json:"result_expires_offset"`
+	ID   string `json:"id"`
+	Name string `json:"name"`
 } //@name TestTypeResult
 
 //CreateTestTypeResult creates a test type result for a specific test type
@@ -2027,19 +2049,15 @@ func (h AdminApisHandler) CreateTestTypeResult(current model.User, group string,
 
 	testTypeID := requestData.TestTypeID
 	name := requestData.Name
-	nextStep := requestData.NextStep
-	nextStepOffset := requestData.NextStepOffset
-	resultExpiresOffset := requestData.ResultExpiresOffset
 
-	testTypeResult, err := h.app.Administration.CreateTestTypeResult(current, group, testTypeID, name, nextStep, nextStepOffset, resultExpiresOffset)
+	testTypeResult, err := h.app.Administration.CreateTestTypeResult(current, group, testTypeID, name, "", nil, nil)
 	if err != nil {
 		log.Println(err.Error())
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	resultItem := createTestTypeResultResponse{ID: testTypeResult.ID, Name: testTypeResult.Name, NextStep: testTypeResult.NextStep,
-		NextStepOffset: testTypeResult.NextStepOffset, ResultExpiresOffset: testTypeResult.ResultExpiresOffset}
+	resultItem := createTestTypeResultResponse{ID: testTypeResult.ID, Name: testTypeResult.Name}
 	data, err = json.Marshal(resultItem)
 	if err != nil {
 		log.Println("Error on marshal a test type result")
@@ -2053,18 +2071,12 @@ func (h AdminApisHandler) CreateTestTypeResult(current model.User, group string,
 }
 
 type updateTestTypeResultRequest struct {
-	Name                string `json:"name" validate:"required"`
-	NextStep            string `json:"next_step" validate:"required"`
-	NextStepOffset      *int   `json:"next_step_offset"`
-	ResultExpiresOffset *int   `json:"result_expires_offset"`
+	Name string `json:"name" validate:"required"`
 } // @name updateTestTypeResultRequest
 
 type updateTestTypeResultResponse struct {
-	ID                  string `json:"id"`
-	Name                string `json:"name"`
-	NextStep            string `json:"next_step"`
-	NextStepOffset      *int   `json:"next_step_offset"`
-	ResultExpiresOffset *int   `json:"result_expires_offset"`
+	ID   string `json:"id"`
+	Name string `json:"name"`
 } // @name TestTypeResult
 
 //UpdateTestTypeResult updates test type result
@@ -2113,19 +2125,15 @@ func (h AdminApisHandler) UpdateTestTypeResult(current model.User, group string,
 	}
 
 	name := requestData.Name
-	nextStep := requestData.NextStep
-	nextStepOffset := requestData.NextStepOffset
-	resultExpiresOffset := requestData.ResultExpiresOffset
 
-	testTypeResult, err := h.app.Administration.UpdateTestTypeResult(current, group, ID, name, nextStep, nextStepOffset, resultExpiresOffset)
+	testTypeResult, err := h.app.Administration.UpdateTestTypeResult(current, group, ID, name, "", nil, nil)
 	if err != nil {
 		log.Println(err.Error())
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	resultItem := updateTestTypeResultResponse{ID: testTypeResult.ID, Name: testTypeResult.Name,
-		NextStep: testTypeResult.NextStep, NextStepOffset: testTypeResult.NextStepOffset, ResultExpiresOffset: testTypeResult.ResultExpiresOffset}
+	resultItem := updateTestTypeResultResponse{ID: testTypeResult.ID, Name: testTypeResult.Name}
 	data, err = json.Marshal(resultItem)
 	if err != nil {
 		log.Println("Error on marshal a test type result")
@@ -2169,11 +2177,8 @@ func (h AdminApisHandler) DeleteTestTypeResult(current model.User, group string,
 }
 
 type getTestTypeResultsResponse struct {
-	ID                  string `json:"id"`
-	Name                string `json:"name"`
-	NextStep            string `json:"next_step"`
-	NextStepOffset      *int   `json:"next_step_offset"`
-	ResultExpiresOffset *int   `json:"result_expires_offset"`
+	ID   string `json:"id"`
+	Name string `json:"name"`
 } // @name TestTypeResult
 
 //GetTestTypeResultsByTestTypeID gets all test type results for a test type
@@ -2206,8 +2211,7 @@ func (h AdminApisHandler) GetTestTypeResultsByTestTypeID(current model.User, gro
 	var resultList []getTestTypeResultsResponse
 	if testTypeResults != nil {
 		for _, item := range testTypeResults {
-			r := getTestTypeResultsResponse{ID: item.ID, Name: item.Name, NextStep: item.NextStep,
-				NextStepOffset: item.NextStepOffset, ResultExpiresOffset: item.ResultExpiresOffset}
+			r := getTestTypeResultsResponse{ID: item.ID, Name: item.Name}
 			resultList = append(resultList, r)
 		}
 	}
@@ -2826,6 +2830,7 @@ type createSymptomRequest struct {
 } //@name createSymptomRequest
 
 //CreateSymptom creates a symptom
+// @Deprecated
 // @Description Creates a symptom
 // @Tags Admin
 // @ID CreateSymptom
@@ -2887,6 +2892,7 @@ type updateSymptomRequest struct {
 } //@name updateSymptomRequest
 
 //UpdateSymptom updates a symptom
+// @Deprecated
 // @Description Updates a symptom.
 // @Tags Admin
 // @ID UpdateSymptom
@@ -2953,6 +2959,7 @@ func (h AdminApisHandler) UpdateSymptom(current model.User, group string, w http
 }
 
 //DeleteSymptom deletes a symptom
+// @Deprecated
 // @Description Deletes a symptom
 // @Tags Admin
 // @ID deleteSymptom
@@ -2990,6 +2997,7 @@ type getSymptomGroupsResponse struct {
 } // @name SymptomGroup
 
 //GetSymptomGroups gets the symptom groups
+// @Deprecated
 // @Description Gives the symptom groups
 // @Tags Admin
 // @ID getSymptomGroups
@@ -3068,6 +3076,7 @@ type symptomRuleItemResponse struct {
 } // @name SymptomRuleItem
 
 //CreateSymptomRule creates a symptom rule
+// @Deprecated
 // @Description Creates a symptom rule.
 // @Tags Admin
 // @ID CreateSymptomRule
@@ -3163,6 +3172,7 @@ type updateSymptomRuleItemRequest struct {
 } //@name updateSymptomRuleItemRequest
 
 //UpdateSymptomRule updates a symptom rule
+// @Deprecated
 // @Description Updates a symptom rule.
 // @Tags Admin
 // @ID UpdateSymptomRule
@@ -3251,6 +3261,7 @@ func (h AdminApisHandler) UpdateSymptomRule(current model.User, group string, w 
 }
 
 //GetSymptomRules gets the symptom rules
+// @Deprecated
 // @Description Gives the symptom rules
 // @Tags Admin
 // @ID getSymptomRules
@@ -3301,6 +3312,7 @@ func (h AdminApisHandler) GetSymptomRules(current model.User, group string, w ht
 }
 
 //DeleteSymptomRule deletes a symptom rule
+// @Deprecated
 // @Description Deletes a symptom rule.
 // @Tags Admin
 // @ID deleteSymptomRule
@@ -3779,6 +3791,205 @@ func (h AdminApisHandler) DeleteAccessRule(current model.User, group string, w h
 	w.Write([]byte("Successfully deleted"))
 }
 
+//GetCRules gets the rules
+// @Description Gives the rules
+// @Tags Admin
+// @ID GetCRules
+// @Accept json
+// @Param county-id query string false "County ID"
+// @Param app-version query string false "App version"
+// @Success 200 {object} string
+// @Security AdminUserAuth
+// @Security AdminGroupAuth
+// @Router /admin/crules [get]
+func (h AdminApisHandler) GetCRules(current model.User, group string, w http.ResponseWriter, r *http.Request) {
+	countyKeys, ok := r.URL.Query()["county-id"]
+	if !ok || len(countyKeys[0]) < 1 {
+		log.Println("url param 'county-id' is missing")
+		return
+	}
+	appVersionKeys, ok := r.URL.Query()["app-version"]
+	if !ok || len(appVersionKeys[0]) < 1 {
+		log.Println("url param 'app-version' is missing")
+		return
+	}
+	countyID := countyKeys[0]
+	appVersion := appVersionKeys[0]
+
+	cRules, err := h.app.Administration.GetCRules(countyID, appVersion)
+	if err != nil {
+		log.Printf("Error on getting crules - %s", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	if cRules == nil {
+		//not found
+		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
+		return
+	}
+
+	data := []byte(cRules.Data)
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(data)
+}
+
+type updateCRulesRequest struct {
+	AppVersion string `json:"app_version" validate:"required"`
+	CountyID   string `json:"county_id" validate:"required"`
+	Data       string `json:"data" validate:"required"`
+} //@name updateCRulesRequest
+
+//UpdateCRules updates the rules
+// @Description Updates the rules.
+// @Tags Admin
+// @ID UpdateCRules
+// @Accept json
+// @Produce json
+// @Param data body updateCRulesRequest true "body data"
+// @Success 200 {object} string
+// @Security AdminUserAuth
+// @Security AdminGroupAuth
+// @Router /admin/crules [put]
+func (h AdminApisHandler) UpdateCRules(current model.User, group string, w http.ResponseWriter, r *http.Request) {
+	bodyData, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		log.Printf("Error on marshal the update crules items  - %s\n", err.Error())
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
+	}
+
+	var requestData updateCRulesRequest
+	err = json.Unmarshal(bodyData, &requestData)
+	if err != nil {
+		log.Printf("Error on unmarshal the update crules items request data - %s\n", err.Error())
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	//validate
+	validate := validator.New()
+	err = validate.Struct(requestData)
+	if err != nil {
+		log.Printf("Error on validating update crules data - %s\n", err.Error())
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	appVersion := requestData.AppVersion
+	countyID := requestData.CountyID
+	data := requestData.Data
+
+	cRules, err := h.app.Administration.UpdateCRules(current, group, countyID, appVersion, data)
+	if err != nil {
+		log.Println(err.Error())
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	resData := []byte(cRules.Data)
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(resData)
+}
+
+//GetSymptoms gets the symptoms
+// @Description Gives the symptoms
+// @Tags Admin
+// @ID GetASymptoms
+// @Accept json
+// @Param app-version query string false "App version"
+// @Success 200 {object} string
+// @Security AdminUserAuth
+// @Security AdminGroupAuth
+// @Router /admin/symptoms [get]
+func (h AdminApisHandler) GetSymptoms(current model.User, group string, w http.ResponseWriter, r *http.Request) {
+	appVersionKeys, ok := r.URL.Query()["app-version"]
+	if !ok || len(appVersionKeys[0]) < 1 {
+		log.Println("url param 'app-version' is missing")
+		return
+	}
+	appVersion := appVersionKeys[0]
+
+	symptoms, err := h.app.Administration.GetSymptoms(appVersion)
+	if err != nil {
+		log.Printf("Error on getting symptoms - %s", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	if symptoms == nil {
+		//not found
+		http.Error(w, http.StatusText(http.StatusNotFound), http.StatusNotFound)
+		return
+	}
+
+	data := []byte(symptoms.Items)
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(data)
+}
+
+type updateSymptomsRequest struct {
+	AppVersion string `json:"app_version" validate:"required"`
+	Items      string `json:"items" validate:"required"`
+} //@name updateSymptomsRequest
+
+//UpdateSymptoms updates the symptoms
+// @Description Updates the symptoms.
+// @Tags Admin
+// @ID UpdateASymptoms
+// @Accept json
+// @Produce json
+// @Param data body updateSymptomsRequest true "body data"
+// @Success 200 {object} string
+// @Security AdminUserAuth
+// @Security AdminGroupAuth
+// @Router /admin/symptoms [put]
+func (h AdminApisHandler) UpdateSymptoms(current model.User, group string, w http.ResponseWriter, r *http.Request) {
+	bodyData, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		log.Printf("Error on marshal the update symptoms items  - %s\n", err.Error())
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
+	}
+
+	var requestData updateSymptomsRequest
+	err = json.Unmarshal(bodyData, &requestData)
+	if err != nil {
+		log.Printf("Error on unmarshal the update symptoms items request data - %s\n", err.Error())
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	//validate
+	validate := validator.New()
+	err = validate.Struct(requestData)
+	if err != nil {
+		log.Printf("Error on validating update symptoms data - %s\n", err.Error())
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	appVersion := requestData.AppVersion
+	items := requestData.Items
+
+	symptoms, err := h.app.Administration.UpdateSymptoms(current, group, appVersion, items)
+	if err != nil {
+		log.Println(err.Error())
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	resData := []byte(symptoms.Items)
+
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+	w.Write(resData)
+}
+
 //GetUserByExternalID gets the user by external id
 // @Description Gets the user by external id.
 // @Tags Admin
@@ -3874,7 +4085,7 @@ func (h ApisHandler) CreateAction(current model.User, group string, w http.Respo
 	encryptedKey := requestData.EncryptedKey
 	encryptedBlob := requestData.EncryptedBlob
 
-	item, err := h.app.Administration.CreateAction(providerID, userID, encryptedKey, encryptedBlob)
+	item, err := h.app.Administration.CreateAction(current, group, providerID, userID, encryptedKey, encryptedBlob)
 	if err != nil {
 		log.Printf("Error on creating an action - %s\n", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)