+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.16.0] - 2020-09-16
+### Added
+- Expose Rules APIs. [#28](https://github.com/rokwire/health-building-block/issues/28)
+
+### Fixed
+- Audit for "action" entity. [#30](https://github.com/rokwire/health-building-block/issues/30)
+
+### Changed
+- Disable test type result validation fields. [#32](https://github.com/rokwire/health-building-block/issues/32)
+
 ## [1.15.0] - 2020-09-10
 ### Added
 - Add order number for the providers tests. [#25](https://github.com/rokwire/health-building-block/issues/25)
@@ -97,7 +107,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update access rule admin api.
 - Delete access rule admin api.
 
-[Unreleased]: https://github.com/rokwire/health-building-block/compare/v1.15.0...HEAD
+[Unreleased]: https://github.com/rokwire/health-building-block/compare/v1.16.0...HEAD
+[1.16.0]: https://github.com/rokwire/health-building-block/compare/v1.15.0...v1.16.0
 [1.15.0]: https://github.com/rokwire/health-building-block/compare/v1.14.0...v1.15.0
 [1.14.0]: https://github.com/rokwire/health-building-block/compare/v1.13.0...v1.14.0
 [1.13.0]: https://github.com/rokwire/health-building-block/compare/v1.12.0...v1.13.0